requirements.txt
Flask>=3.0,<4.0
gunicorn>=21.2,<22.0
python-docx>=1.1.0,<2.0
python-pptx>=0.6.23,<1.0
openpyxl>=3.1.2,<4.0
olefile>=0.47,<1.0
PyPDF2>=3.0.1,<4.0
